Refactors Zoom component
